+++ connector-packager/connector_packager/jar_jdk_packager.py
@@ -65,6 +65,17 @@ def get_min_support_version(file_list: List[ConnectorFile], cur_min_version_tabl
             elif (oauthConfigId is not None and 2021.4 > float(min_version_tableau)):
                 min_version_tableau = "2021.4"
                 reasons.append("Connector contains an oauthConfigId field, which was added in the 2021.4 release")
+            instanceUrlSuffix = pluginOAuthConfigRoot.find('.//instanceUrlSuffix')
+            configLabel = pluginOAuthConfigRoot.find('.//configLabel')
+            if (configLabel is not None and 2023.2 > float(min_version_tableau)):
+                min_version_tableau = "2023.2"
+                reasons.append("Connector uses configLabel field, which was added in the 2023.2 release")
+            elif (instanceUrlSuffix is not None and 2023.1 > float(min_version_tableau)):
+                min_version_tableau = "2023.1"
+                reasons.append("Connector uses instanceUrlSuffix field, which was added in the 2023.1 release")
+            elif (oauthConfigId is not None and 2021.4 > float(min_version_tableau)):
+                min_version_tableau = "2021.4"
+                reasons.append("Connector contains an oauthConfigId field, which was added in the 2021.4 release")
             elif 2021.1 > float(min_version_tableau):
                 min_version_tableau = "2021.1"
             reasons.append("Connector uses OAuth, which was added in the 2021.1 release")
@@ -80,8 +91,14 @@ def get_min_support_version(file_list: List[ConnectorFile], cur_min_version_tabl
         elif connector_file.file_type == "manifest":
             manifestRoot = ET.parse(input_dir / connector_file.file_name).getroot()
             oauthConfigs = manifestRoot.findall('.//oauth-config')
-            if (oauthConfigs is not None and len(oauthConfigs) > 1 and 2021.4 > float(min_version_tableau)):
-                min_version_tableau = "2021.4"
+            if (oauthConfigs is not None and len(oauthConfigs) > 1 and 2023.1 > float(min_version_tableau)):
+                min_version_tableau = "2023.1"
+                reasons.append("Support for multiple OAuth configs was added in the 2023.1 release")
+            elif (oauthConfigs is not None and len(oauthConfigs) == 1):
+                firstConfig = oauthConfigs[0]
+                if firstConfig.attrib['file'] == "null_config" and 2024.1 > float(min_version_tableau):
+                    min_version_tableau = "2024.1"
+                    reasons.append("Connector uses Null OAuth Config, which was added in the 2024.1 release")
 
     if version.parse(cur_min_version_tableau) > version.parse(min_version_tableau):
         reasons.append("min-tableau-version set to " + cur_min_version_tableau + ", since that is higher than calculated version of " + min_version_tableau)