+++ connector-packager/tests/test_xml_parser.py
@@ -13,6 +13,7 @@ class TestXMLParser(unittest.TestCase):
 
         # Test valid connector
         expected_class_name = "postgres_odbc"
+        expected_connector_version = "0.0.0"
         expected_file_list = [
             ConnectorFile("manifest.xml", "manifest"),
             ConnectorFile("connection-dialog.tcd", "connection-dialog"),
@@ -21,7 +22,7 @@ class TestXMLParser(unittest.TestCase):
             ConnectorFile("connectionResolver.tdr", "connection-resolver"),
             ConnectorFile("resources-en_US.xml", "resource")]
 
-        actual_file_list, actual_class_name = self.parser_test_case(TEST_FOLDER / Path("valid_connector"),
+        actual_file_list, actual_class_name, actual_connector_version = self.parser_test_case(TEST_FOLDER / Path("valid_connector"),
                                                                     expected_file_list, expected_class_name)
 
         self.assertTrue(actual_file_list, "Valid connector did not return a file list")
@@ -29,25 +30,27 @@ class TestXMLParser(unittest.TestCase):
                         "Actual file list does not match expected for valid connector")
         self.assertTrue(actual_class_name == expected_class_name,
                         "Actual class name does not match expected for valid connector")
+        self.assertTrue(actual_connector_version == expected_connector_version,
+                        "Actual connector version does not match expected for valid connector")
 
         print("\nTest invalid connector. Throws XML validation error.")
 
-        actual_file_list, actual_class_name = self.parser_test_case(TEST_FOLDER / Path("broken_xml"),
+        actual_file_list, actual_class_name, actual_connector_version = self.parser_test_case(TEST_FOLDER / Path("broken_xml"),
                                                                     expected_file_list, expected_class_name)
         self.assertFalse(actual_file_list, "Invalid connector returned a file list when it should not have")
 
         print("\nTest connector with class name mismatch. Throws XML validation error.")
-        actual_file_list, actual_class_name = self.parser_test_case(TEST_FOLDER / Path("wrong_class"),
+        actual_file_list, actual_class_name, actual_connector_version = self.parser_test_case(TEST_FOLDER / Path("wrong_class"),
                                                                     expected_file_list, expected_class_name)
         self.assertFalse(actual_file_list, "Connector with class name mismatch returned a file list when it shouldn't")
 
         # Test connector with non-https url
-        actual_file_list, actual_class_name = self.parser_test_case(TEST_FOLDER / Path("non_https"),
+        actual_file_list, actual_class_name, actual_connector_version = self.parser_test_case(TEST_FOLDER / Path("non_https"),
                                                                     expected_file_list, expected_class_name)
         self.assertFalse(actual_file_list, "Connector with non-https urls returned a file list when it shouldn't")
-
-        # Test connector with missing English translation
-        actual_file_list, actual_class_name = self.parser_test_case(TEST_FOLDER / Path("missing_english_translation"),
+        
+        # Test connector with missing English transaltion
+        actual_file_list, actual_class_name, actual_connector_version = self.parser_test_case(TEST_FOLDER / Path("missing_english_translation"),
                                                                     expected_file_list, expected_class_name)
         self.assertFalse(actual_file_list, "Connector with localized strings but without a resources-en_US.xml file "
                                            "returned a file list when it shouldn't")
@@ -55,6 +58,7 @@ class TestXMLParser(unittest.TestCase):
     def test_generate_file_list_mcd(self):
         # Test modular dialog connector
         expected_class_name = "postgres_mcd"
+        expected_connector_version = "0.0.1"
         expected_file_list = [
             ConnectorFile("manifest.xml", "manifest"),
             ConnectorFile("connectionFields.xml", "connection-fields"),
@@ -64,7 +68,7 @@ class TestXMLParser(unittest.TestCase):
             ConnectorFile("connectionResolver.xml", "connection-resolver"),
             ConnectorFile("connectionProperties.js", "script")]
 
-        actual_file_list, actual_class_name = self.parser_test_case(TEST_FOLDER / Path("modular_dialog_connector"),
+        actual_file_list, actual_class_name, actual_connector_version = self.parser_test_case(TEST_FOLDER / Path("modular_dialog_connector"),
                                                                     expected_file_list, expected_class_name)
 
         self.assertTrue(actual_file_list, "Valid connector did not return a file list")
@@ -72,10 +76,13 @@ class TestXMLParser(unittest.TestCase):
                         "Actual file list does not match expected for valid connector")
         self.assertTrue(actual_class_name == expected_class_name,
                         "Actual class name does not match expected for valid connector")
+        self.assertTrue(actual_connector_version == expected_connector_version,
+                        "Actual connector version does not match expected for valid connector")
 
     def test_generate_file_list_oauth(self):
         # Test oauth connector
         expected_class_name = "test_oauth"
+        expected_connector_version = "0.0.1"
         expected_file_list = [
             ConnectorFile("manifest.xml", "manifest"),
             ConnectorFile("connectionFields.xml", "connection-fields"),
@@ -86,7 +93,7 @@ class TestXMLParser(unittest.TestCase):
             ConnectorFile("connectionResolver.xml", "connection-resolver"),
             ConnectorFile("connectionProperties.js", "script")]
 
-        actual_file_list, actual_class_name = self.parser_test_case(TEST_FOLDER / Path("oauth_connector"),
+        actual_file_list, actual_class_name, actual_connector_version = self.parser_test_case(TEST_FOLDER / Path("oauth_connector"),
                                                                     expected_file_list, expected_class_name)
 
         self.assertTrue(actual_file_list, "Valid connector did not return a file list")
@@ -94,6 +101,8 @@ class TestXMLParser(unittest.TestCase):
                         "Actual file list does not match expected for valid connector")
         self.assertTrue(actual_class_name == expected_class_name,
                         "Actual class name does not match expected for valid connector")
+        self.assertTrue(actual_connector_version == expected_connector_version,
+                        "Actual connector version does not match expected for valid connector")
 
     def parser_test_case(self, test_folder, expected_file_list, expected_class_name):
 
@@ -101,5 +110,6 @@ class TestXMLParser(unittest.TestCase):
 
         actual_file_list = xml_parser.generate_file_list()
         actual_class_name = xml_parser.class_name
+        actual_connector_version = xml_parser.connector_version
 
-        return actual_file_list, actual_class_name
+        return actual_file_list, actual_class_name, actual_connector_version